runtime: get the disk name from the sysfs path
The naming of the disk partitions in /dev varies among the types of
block devices. For SCSI disks, it could be /dev/sda1. For NVME disks, it
could be /dev/nvme0n1p1. To get the disk name reliably, we check the
real path of the partition device in /sys/class/block.

For example, to look up the disk device name of /dev/nvme0n1p1, we first
get the link of /sys/class/block/nvme0n1p1:

../../devices/pci0000:00/0000:00:06.0/0000:02:00.0/nvme/nvme0/nvme0n1/nvme0n1p1

Then we retrieve the disk name from the upper level directory: nvme0n1
and return "/dev/nvme0n1" for the further operation, e.g. the inspection
of the GPT table.

Signed-off-by: Gary Lin <glin@suse.com>