efi-variable: matching the certificate in the key databases
When verifying the TPM2_EFI_VARIABLE_AUTHORITY event, the event data may
not cover the whole EFI variable. For example, edk2 only measures the
x509 certificate in "db" that verifies the EFI image, not the whole "db"
variable. Instead of rehashing the EFI variable, just check if the raw
event data is from the given EFI variable.

(_GNU_SOURCE is introduced to support memem().)

Fixes: https://github.com/okirch/pcr-oracle/issues/9

Signed-off-by: Gary Lin <glin@suse.com>