Try to handle synthetic Shim variable
This is an event emitted by the shim loader. There is no "Shim" variable, this is
a fake variable. The value is the X509 certificate used to verify the authenticode
signature on the next stage loader (usually grub).

We try to predict this certificate by looking at the cert table of the next
stage loader, and extracting the certificate of the signing authority from it.

This is a bit flaky, as a PECOFF image may contain several  authenticode
signatures - and only one has to match. So there is potential for confusion,
but it's probably more theoretical than practical.

Signed-off-by: Olaf Kirch <okir@suse.de>