Handle NO_ACTION events when predicting PCR values
Per "TCG PC Client Platform Firmware Profile Specification", there
are 3 types of NO_ACTION events:

1. Specification ID event
2. Reference Manifest event
3. Startup Locality event

The first two types are not measured so those events must be skipped
when predicting PCR values. On the other hand, the Startup Locality
event may change the initial PCR0 value. A Startup Locality event is
always the first event and for PCR0, and the last byte of PCR0 has
to be set to the locality value.

Fixes: https://github.com/okirch/pcr-oracle/issues/13

Signed-off-by: Gary Lin <glin@suse.com>