Fix ESYS_TR_RH_* constant mess for older libtss2 versions
Commit fa31276ccd14a275fa0f94341cfbb1cc8157fd3d changed the hierarchy
values for several Esys calls from TPM2_RH_{NULL,OWNER} to ESYS_TR_RH_*
to reflect the upstream API change.

Unfortunately, while the code still compiles against older libtss2
versions, it won't run there.

This patch introduces a run-time check for the libtss2 library version
and selects the right set of constants based on what we find.

It's ugly as hell, but for the time being, it's the best I could come
up with in finite time.

Signed-off-by: Olaf Kirch <okir@suse.com>